+++ spec/request_spec.rb
@@ -249,6 +249,11 @@ describe RestClient::Request do
       lambda { @request.process_result(res) }.should raise_error(RestClient::Redirect) { |e| e.url.should == 'http://some/index' }
     end
 
+    it "handles redirects with relative path and query string" do
+      res = mock('response', :code => '301', :header => { 'Location' => 'index?q=1' }, :[] => ['content-encoding' => ''], :body => '' )
+      lambda { @request.process_result(res) }.should raise_error(RestClient::Redirect) { |e| e.url.should == 'http://some/index?q=1' }
+    end
+
     it "handles redirects with absolute paths" do
       @request.instance_variable_set('@url', 'http://some/place/else')
       res = mock('response', :code => '301', :header => { 'Location' => '/index' }, :[] => ['content-encoding' => ''], :body => '' )